+++ ktlint-ruleset-experimental/src/main/kotlin/com/pinterest/ktlint/ruleset/experimental/ExperimentalRuleSetProvider.kt
@@ -15,6 +15,7 @@ class ExperimentalRuleSetProvider : RuleSetProvider {
         SpacingAroundDoubleColonRule(),
         SpacingBetweenDeclarationsWithCommentsRule(),
         SpacingBetweenDeclarationsWithAnnotationsRule(),
-        SpacingAroundAngleBracketsRule()
+        SpacingAroundAngleBracketsRule(),
+        SpacingAroundUnaryOperatorRule()
     )
 }
