+++ ktlint-ruleset-experimental/src/test/kotlin/com/pinterest/ktlint/ruleset/experimental/SpacingAroundUnaryOperatorRuleTest.kt
@@ -1,22 +1,22 @@
-package com.pinterest.ktlint.ruleset.standard
+package com.pinterest.ktlint.ruleset.experimental
 
 import com.pinterest.ktlint.test.diffFileFormat
 import com.pinterest.ktlint.test.diffFileLint
 import org.assertj.core.api.Assertions.assertThat
 import org.junit.Test
 
-class SpacingInUnaryExpressionRuleTest {
+class SpacingAroundUnaryOperatorRuleTest {
 
     @Test
     fun testLint() {
-        assertThat(SpacingInUnaryExpressionRule().diffFileLint("spec/unary-op-spacing/lint.kt.spec"))
+        assertThat(SpacingAroundUnaryOperatorRule().diffFileLint("spec/unary-op-spacing/lint.kt.spec"))
             .isEmpty()
     }
 
     @Test
     fun testFormat() {
         assertThat(
-            SpacingInUnaryExpressionRule().diffFileFormat(
+            SpacingAroundUnaryOperatorRule().diffFileFormat(
                 srcPath = "spec/unary-op-spacing/format.kt.spec",
                 expectedPath = "spec/unary-op-spacing/format-expected.kt.spec"
             )
