+++ ktlint-ruleset-experimental/src/test/resources/spec/unary-op-spacing/format.kt.spec
