+++ ktlint-ruleset-experimental/src/test/resources/spec/unary-op-spacing/lint.kt.spec
